requirements.txt
numpy==1.25.2
pandas==2.0.3
Flask==2.3.3
openpyxl==3.1.2
google-api-python-client==2.89.0
google-auth==2.23.0
google-auth-httplib2==0.1.0
httplib2==0.22.0
requests==2.31.0
